Bait changes, /daily & misc fixes

+++ core/src/main/resources/core-migrations/global/V1__Initial_global_core.sql
@@ -94,7 +94,7 @@ INSERT IGNORE INTO property_map VALUES ("LAST_LOGIN", "long");
 INSERT IGNORE INTO property_map VALUES ("NO_DOMINANCE_COOLDOWN", "long");
 INSERT IGNORE INTO property_map VALUES ("LAST_TNTED", "long");
 INSERT IGNORE INTO property_map VALUES ("TIME_CREATED", "long");
-INSERT IGNORE INTO property_map VALUES ("EXPERIENCE", "long");
+INSERT IGNORE INTO property_map VALUES ("EXPERIENCE", "double");
 INSERT IGNORE INTO property_map VALUES ("LUNAR", "boolean");
 INSERT IGNORE INTO property_map VALUES ("TIME_PLAYED", "long");
